+++ Cargo.toml
@@ -2,7 +2,7 @@
 resolver = "2"
 members = [
   "crates/l2/starknet-core-contract-client",
-  "crates/l2/ethereum-instance", "crates/l2/starknet-eth-bridge-client", "crates/l2/starknet-token-bridge-client", "crates/l2/starkgate-manager-client", "crates/l2/utils", "crates/l2/starknet-erc20-client", "crates/l2/starkgate-registry-client", "crates/l2/starknet-proxy-client", "crates/l3/appchain-core-contract-client",  "crates/l3/starknet-instance", "crates/l3/common"
+  "crates/l2/ethereum-instance", "crates/l2/starknet-eth-bridge-client", "crates/l2/starknet-token-bridge-client", "crates/l2/starkgate-manager-client", "crates/l2/utils", "crates/l2/starknet-erc20-client", "crates/l2/starkgate-registry-client", "crates/l2/starknet-proxy-client", "crates/l3/appchain-core-contract-client",  "crates/l3/starknet-instance", "crates/l3/appchain-utils"
 ]
 
 [workspace.package]
