+++ crates/l3/appchain-core-contract-client/Cargo.toml
@@ -10,7 +10,7 @@ dirs = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
 hex = { workspace = true }
-common = { path = "../common" }
+appchain-utils = { path = "../appchain-utils" }
 tokio = { workspace = true }
 log = { workspace = true }
 num-traits = { workspace = true }
